code/requirements.txt
pydicom
numpy
pillow
pandas
boto3
mxnet
opencv-python
